Now using R-based permutations instead of Hail to maximize speed and avoid Apache recursion errors

+++ scripts/permute/_gene_permute.sh
@@ -18,6 +18,7 @@ source utils/hail_utils.sh
 
 readonly spark_dir="data/tmp/spark"
 readonly hail_script="scripts/permute/_gene_permute.py"
+readonly rscript="scripts/permute/_gene_permute.R"
 
 readonly chr=${1?Error: Missing arg1 (phenotype)}
 readonly input_path=${2?Error: Missing arg1 (phenotype)}
@@ -36,27 +37,42 @@ readonly checkpoint="${out_prefix_gene}_${sge_seed}_checkpoint.mt"
 readonly input_path_gene=$(echo ${input_path} | sed -e "s/GENE/${gene}/g")
 
 
-SECONDS=0
-set_up_hail
-set_up_pythonpath_legacy
-python3 ${hail_script} \
+set_up_rpy
+Rscript ${rscript} \
   --chrom "chr${chr}" \
   --input_path ${input_path_gene} \
-  --input_type ${input_type} \
-  --replicates ${replicates} \
+  --permutations ${replicates} \
   --out_prefix ${out_prefix_gene} \
-  --out_type ${out_type} \
+  --vcf_id ${gene} \
   --seed ${sge_seed} \
   && print_update "Finished permuting phase for chr${chr}" ${SECONDS} \
   || raise_error "Permuting phase for chr${chr} failed"
- #--gene ${gene} \
-  #--checkpoint ${checkpoint} \
-  
-rm -rf ${checkpoint}
+  module purge
+  module load BCFtools/1.12-GCC-10.3.0
+  bgzip "${out_prefix_gene}.vcf"
+  make_tabix "${out_prefix_gene}.vcf.gz" "csi"
 
-module purge
-module load BCFtools/1.12-GCC-10.3.0
-make_tabix "${out_prefix_gene}.vcf.bgz" "csi"
-
+#if [ ! -f "${out_prefix_gene}.vcf.bgz" ]; then
+if [ 1 -eq 2 ]; then
+  SECONDS=0
+  set_up_hail
+  set_up_pythonpath_legacy
+  python3 ${hail_script} \
+    --chrom "chr${chr}" \
+    --input_path ${input_path_gene} \
+    --input_type ${input_type} \
+    --replicates ${replicates} \
+    --out_prefix ${out_prefix_gene} \
+    --out_type ${out_type} \
+    --seed ${sge_seed} \
+    && print_update "Finished permuting phase for chr${chr}" ${SECONDS} \
+    || raise_error "Permuting phase for chr${chr} failed"
+  rm -rf ${checkpoint}
+  module purge
+  module load BCFtools/1.12-GCC-10.3.0
+  make_tabix "${out_prefix_gene}.vcf.bgz" "csi"
+else
+  >&2 echo "${out_prefix_gene}.vcf.bgz already exists. Skipping.."
+fi
